+++ Assets/HeroesFlight/System/Gameplay/Controllers/Projectiles/ProjectileControllerBase.cs
@@ -10,7 +10,7 @@ namespace HeroesFlightProject.System.Gameplay.Controllers
         Transform view;
         public event Action OnEnded;
         Vector2 currentDirection = default;
-        int damage;
+        float damage;
         float currentLifetime;
 
 
@@ -31,7 +31,7 @@ namespace HeroesFlightProject.System.Gameplay.Controllers
             Destroy(this.gameObject);
         }
 
-        public void SetupProjectile(int targetDamage, Transform currentTarget, Vector2 targetDirection)
+        public void SetupProjectile(float targetDamage, Transform currentTarget, Vector2 targetDirection)
         {
             currentDirection = targetDirection;
             damage = targetDamage;
